feat: working claude example
