--- typescript/examples/model-context-protocol-cdp-chatbot/.env-local
@@ -1,3 +1,0 @@
-OPENAI_API_KEY=
-CDP_API_KEY_NAME=
-CDP_API_KEY_PRIVATE_KEY=
