+++ typescript/examples/model-context-protocol-cdp-chatbot/claude_desktop_config-local.json
@@ -1,0 +1,14 @@
+{
+    "mcpServers": {
+        "agentkit": {
+            "command": "node",
+            "args": [
+                "/Users/carsonroscoe/Documents/Github/Coinbase/agentkit/typescript/examples/model-context-protocol-cdp-chatbot/build/index.js"
+            ],
+            "env": {
+                "CDP_API_KEY_NAME": "YOUR_CDP_API_KEY_NAME",
+                "CDP_API_KEY_PRIVATE_KEY": "YOUR_CDP_API_KEY_PRIVATE_KEY"
+            }
+        }
+    }
+}
