+++ typescript/examples/model-context-protocol-cdp-chatbot/package.json
@@ -5,9 +5,11 @@
   "private": true,
   "author": "Coinbase Inc.",
   "license": "Apache-2.0",
+  "bin": {
+    "agentkit": "./build/index.js"
+  },
   "scripts": {
-    "start": "NODE_OPTIONS='--no-warnings' ts-node ./chatbot.ts",
-    "dev": "nodemon ./chatbot.ts",
+    "build": "tsc && node -e \"require('fs').chmodSync('build/index.js', '755')\"",
     "lint": "eslint -c .eslintrc.json *.ts",
     "lint:fix": "eslint -c .eslintrc.json *.ts --fix",
     "format": "prettier --write \"**/*.{ts,js,cjs,json,md}\"",
