+++ typescript/examples/model-context-protocol-cdp-chatbot/tsconfig.json
@@ -2,8 +2,10 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "preserveSymlinks": true,
-    "outDir": "./dist",
-    "rootDir": "."
+    "outDir": "./build",
+    "rootDir": ".",
+    "module": "Node16",
+    "moduleResolution": "Node16"
   },
   "include": ["*.ts"]
 }
